+++ src/codegen/generate-js2native.ts
@@ -116,24 +116,25 @@ export function getJS2NativeCPP() {
 
   return [
     `#pragma once`,
+    `#include "root.h"`,
     ...files.map(filename => `#include ${JSON.stringify(filename)}`),
-    "namespace JS2NativeGenerated {",
+    "\n" + "namespace JS2NativeGenerated {",
     "using namespace Bun;",
     "using namespace JSC;",
-    "using namespace WebCore;",
+    "using namespace WebCore;" + "\n",
     ...nativeCalls
       .filter(x => x.type === "zig")
       .flatMap(call => [
-        `extern JSC_CALLCONV JSC::EncodedJSValue ${symbol(call)}_workaround(Zig::GlobalObject*);`,
+        `extern "C" SYSV_ABI JSC::EncodedJSValue ${symbol(call)}_workaround(Zig::GlobalObject*);` + "\n",
         `JSC::JSValue ${symbol(call)}(Zig::GlobalObject* global) {`,
         `  return JSValue::decode(${symbol(call)}_workaround(global));`,
-        `}`,
+        `}` + "\n\n",
       ]),
     ...wrapperCalls.map(x => {
       if (x.wrap_kind === "new-function") {
         return [
           x.type === "zig"
-            ? `extern JSC_CALLCONV JSC_DECLARE_HOST_FUNCTION(${symbol({
+            ? `BUN_DECLARE_HOST_FUNCTION(${symbol({
                 type: "zig",
                 symbol: x.symbol_taget,
               })});`
@@ -158,7 +159,7 @@ export function getJS2NativeCPP() {
 
 export function getJS2NativeZig(gs2NativeZigPath: string) {
   return [
-    "//! This file is generated by src/codegen/js2native-generator.ts based on seen calls to the $zig() JS macro",
+    "//! This file is generated by src/codegen/generate-js2native.ts based on seen calls to the $zig() JS macro",
     `const JSC = @import("root").bun.JSC;`,
     ...nativeCalls
       .filter(x => x.type === "zig")
